rename IsRequested to IsRequestedOption
